Move admin functions to adminController
