+++ src/main/java/com/piper/valley/models/repository/StoreRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface StoreRepository extends JpaRepository<Store, Long> {
-	Optional<Store> findOneById(long id);
+	
 }
